Fix codacy alert on Shared project.
