+++ src/AlmaIt.dotnet.Heroes.Shared/Enumeration/ComicBookStatus.cs
@@ -1,9 +1,23 @@
 namespace AlmaIt.dotnet.Heroes.Shared.Enumeration
 {
+    /// <summary>
+    /// Enum for comic book status.
+    /// </summary>
     public enum ComicBookStatus
     {
+        /// <summary>
+        /// Collection status.
+        /// </summary>
         Collection = 0,
+
+        /// <summary>
+        /// Whishlist status.
+        /// </summary>
         Wishlist = 1,
-        None = 2
+
+        /// <summary>
+        /// None status.
+        /// </summary>
+        None = 2,
     }
 }
