+++ src/AlmaIt.dotnet.Heroes.Shared/Models/IDataObject.cs
@@ -1,17 +1,13 @@
-using System;
-
 namespace AlmaIt.dotnet.Heroes.Shared.Models
 {
-    /// <summary>
-    /// Base interface defintion of an object model.
-    /// </summary>
-    /// <typeparam name="TIdentity">primary key type</typeparam>
+    using System;
+
+    /// <summary>Base interface defintion of an object model.</summary>
+    /// <typeparam name="TIdentity">Unique identifier (primary key).</typeparam>
     public interface IDataObject<TIdentity>
         where TIdentity : IComparable<TIdentity>
     {
-        /// <summary>
-        /// Get or Set primary key
-        /// </summary>
+        /// <summary>Gets or sets the primary key.</summary>
         TIdentity Id { get; set; }
     }
 }
