--- src/AlmaIt.dotnet.Heroes.Shared/Models/SortingParams.cs
@@ -1,7 +1,0 @@
-namespace AlmaIt.dotnet.Heroes.Shared.Models
-{
-    public class SortingParams
-    {
-        public string SortBy { get; set; } = "";
-    }
-}
