+++ src/AlmaIt.dotnet.Heroes.Shared/stylecop.json
@@ -1,0 +1,22 @@
+﻿{
+  // ACTION REQUIRED: This file was automatically added to your project, but it
+  // will not take effect until additional steps are taken to enable it. See the
+  // following page for additional information:
+  //
+  // https://github.com/DotNetAnalyzers/StyleCopAnalyzers/blob/master/documentation/EnableConfiguration.md
+
+  "$schema": "https://raw.githubusercontent.com/DotNetAnalyzers/StyleCopAnalyzers/master/StyleCop.Analyzers/StyleCop.Analyzers/Settings/stylecop.schema.json",
+  "settings": {
+    "documentationRules": {
+      "companyName": "Almait",
+      "copyrightText": "Property of Almait",
+      "documentPrivateElements": false,
+      "documentPrivateFields": false,
+      "xmlHeader": false,
+      "documentationCulture": "en-US",
+      "documentInterfaces": true,
+      "documentExposedElements": true,
+      "documentInternalElements": true
+    }
+  }
+}
